+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 Sunflower Framework
 ======================================
+1.2.1 - 
+
+  - add injector processor
+  - rename Server to AbstractServer
+  
 1.2.0 - 2017-10-21
   
   - add sf basic archetypes
@@ -10,7 +15,7 @@ Sunflower Framework
 1.1.0 - 2017-10-18
 
   - sf-sec move to sf-undertow
-  - sf-undertow become generic http server
+  - sf-undertow become generic http Server
 
 1.0.0 - 2017-10-15